agent-backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
python-dotenv>=1.0.0
SQLAlchemy>=2.0.30
psycopg2-binary>=2.9.9
PyJWT>=2.9.0
pydantic>=2.7.0
groq>=0.20.0
langchain-groq>=0.1.9
langchain-core>=0.3.0
langgraph>=0.2.0
httpx>=0.27.0
pinecone-client>=5.0.1
sentence-transformers>=3.0.1
langchain-huggingface>=0.0.3
sse-starlette>=2.1.3
